// apps/api/src/v1/users/user-quote/types/quote-service-key.ts
export type QuoteServiceKey =
  | "userSignUp"
  | "waitlist"
  | "landingPage"
  | "emailsSent"
  | "emailsTemplates"
  | "domains.email"
  | "domains.general";

export const QUOTE_SERVICE_KEYS: QuoteServiceKey[] = [
  "userSignUp",
  "waitlist",
  "landingPage",
  "emailsSent",
  "emailsTemplates",
  "domains.email",
  "domains.general",
];